AF-243 Add get user by ID test coverage
Correctly populate the UserResponse object to display
the user's first name, last name, and email.
